+++ src/config.mjs
@@ -1,5 +1,6 @@
 const config = {
   StorageFileName: 'storage.json',
+  CodeLength: 5,
 };
 
 export default config;