+++ src/cutURL.mjs
@@ -1,11 +1,12 @@
-const cutURL = (req, res) => {
-  const urlRegex = /^(https?:\/\/)?[\d\w]+\.[\w]+(\/.*)*/;
+import ShortCode from './data/ShortCode.mjs';
+import URLTester from './data/URLTester.mjs';
 
-  if (req.body.url === undefined || !urlRegex.test(req.body.url)) {
+const cutURL = (req, res) => {
+  if (req.body.url === undefined || !URLTester.isValid(req.body.url)) {
     return res.status(400).send('Bad request');
   }
 
-  const code = 'xxxxx'.replace(/x/g, () => Math.floor(Math.random() * 16).toString(16));
+  const code = ShortCode.generate();
 
   console.log('cutURL /%s from %s', code, req.body.url);