+++ test/config.test.mjs
@@ -3,3 +3,7 @@ import config from '../src/config.mjs';
 it('config/StorageFileName/+', () => {
   expect(config.StorageFileName).toBeTruthy();
 });
+
+it('config/CodeLength/+', () => {
+  expect(config.CodeLength).toBeTruthy();
+});