+++ CHANGELOG.md
@@ -6,3 +6,4 @@
   - `#1`: As Web Dev, I want to have an API to create shorten URL.
 - **Refactor**
   - `#9`: As dev, i want to refactor data into app context.
+  - `#3`: As dev, I want to refactor `cutURL.mjs` and test with jest.